requirements.txt
# Core dependencies
streamlit>=1.32.0
anthropic>=0.21.0
python-dotenv>=1.0.0
pandas>=2.2.0
pyyaml>=6.0.1

# Scheduling and optimization
ortools>=9.9.0

# Data handling and visualization
plotly>=5.19.0
faker>=24.0.0

# Utilities
python-dateutil>=2.9.0
